# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeuronDrive API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Vehicle {
  id: number;
  vin: string;
  make: string;
  model: string;
  year: number;
  engineCode?: string;
  mileageKm: number;
  healthScore: number;
  createdAt: string;
}

export interface CreateVehicleBody {
  vin: string;
  make: string;
  model: string;
  year: number;
  engineCode?: string;
  mileageKm: number;
}

export interface Session {
  id: number;
  vehicleId: number;
  startedAt: string;
  endedAt?: string | null;
  maxRpm?: number | null;
  maxSpeedKmh?: number | null;
  avgHealthScore?: number | null;
  distanceKm?: number | null;
  anomalyDetected: boolean;
}

export interface CreateSessionBody {
  vehicleId: number;
  maxRpm?: number;
  maxSpeedKmh?: number;
  avgHealthScore?: number;
  distanceKm?: number;
  anomalyDetected: boolean;
}

export interface SessionSummary {
  totalSessions: number;
  avgHealthScore: number;
  totalDistanceKm: number;
  anomalyCount: number;
  recentSessions: Session[];
}

export type DiagnosticSeverity =
  (typeof DiagnosticSeverity)[keyof typeof DiagnosticSeverity];

export const DiagnosticSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Diagnostic {
  id: number;
  vehicleId?: number | null;
  dtcCode: string;
  severity: DiagnosticSeverity;
  description: string;
  rootCauses: string[];
  repairSteps: string[];
  estimatedCostMin?: number | null;
  estimatedCostMax?: number | null;
  resolvedAt?: string | null;
  createdAt: string;
}

export interface AnalyzeDtcBody {
  dtcCode: string;
  vehicleId?: number;
}

export type PredictionStatus =
  (typeof PredictionStatus)[keyof typeof PredictionStatus];

export const PredictionStatus = {
  good: "good",
  watch: "watch",
  warning: "warning",
  critical: "critical",
} as const;

export interface Prediction {
  id: number;
  vehicleId: number;
  component: string;
  status: PredictionStatus;
  kmUntilService: number;
  confidence: number;
  lastCheckedAt: string;
}

export interface EcuMap {
  id: number;
  vin: string;
  make: string;
  model: string;
  year: number;
  calId: string;
  description?: string | null;
  checksumHex?: string | null;
  sizeBytes?: number | null;
  createdAt: string;
}

export interface CreateMapBody {
  vin: string;
  make: string;
  model: string;
  year: number;
  calId: string;
  description?: string;
  checksumHex?: string;
  sizeBytes?: number;
}

export type ListSessionsParams = {
  vehicleId?: number;
};

export type ListDiagnosticsParams = {
  vehicleId?: number;
};

export type ListPredictionsParams = {
  vehicleId: number;
};