# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeuronDrive API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all registered vehicles
 */
export const ListVehiclesResponseItem = zod.object({
  id: zod.number(),
  vin: zod.string(),
  make: zod.string(),
  model: zod.string(),
  year: zod.number(),
  engineCode: zod.string().optional(),
  mileageKm: zod.number(),
  healthScore: zod.number(),
  createdAt: zod.coerce.date(),
});
export const ListVehiclesResponse = zod.array(ListVehiclesResponseItem);

/**
 * @summary Register a new vehicle
 */
export const CreateVehicleBody = zod.object({
  vin: zod.string(),
  make: zod.string(),
  model: zod.string(),
  year: zod.number(),
  engineCode: zod.string().optional(),
  mileageKm: zod.number(),
});

/**
 * @summary Get a vehicle by ID
 */
export const GetVehicleParams = zod.object({
  id: zod.coerce.number(),
});

export const GetVehicleResponse = zod.object({
  id: zod.number(),
  vin: zod.string(),
  make: zod.string(),
  model: zod.string(),
  year: zod.number(),
  engineCode: zod.string().optional(),
  mileageKm: zod.number(),
  healthScore: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a vehicle
 */
export const DeleteVehicleParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all OBD sessions
 */
export const ListSessionsQueryParams = zod.object({
  vehicleId: zod.coerce.number().optional(),
});

export const ListSessionsResponseItem = zod.object({
  id: zod.number(),
  vehicleId: zod.number(),
  startedAt: zod.coerce.date(),
  endedAt: zod.coerce.date().nullish(),
  maxRpm: zod.number().nullish(),
  maxSpeedKmh: zod.number().nullish(),
  avgHealthScore: zod.number().nullish(),
  distanceKm: zod.number().nullish(),
  anomalyDetected: zod.boolean(),
});
export const ListSessionsResponse = zod.array(ListSessionsResponseItem);

/**
 * @summary Create a new OBD session
 */
export const CreateSessionBody = zod.object({
  vehicleId: zod.number(),
  maxRpm: zod.number().optional(),
  maxSpeedKmh: zod.number().optional(),
  avgHealthScore: zod.number().optional(),
  distanceKm: zod.number().optional(),
  anomalyDetected: zod.boolean(),
});

/**
 * @summary Get session by ID
 */
export const GetSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetSessionResponse = zod.object({
  id: zod.number(),
  vehicleId: zod.number(),
  startedAt: zod.coerce.date(),
  endedAt: zod.coerce.date().nullish(),
  maxRpm: zod.number().nullish(),
  maxSpeedKmh: zod.number().nullish(),
  avgHealthScore: zod.number().nullish(),
  distanceKm: zod.number().nullish(),
  anomalyDetected: zod.boolean(),
});

/**
 * @summary Get aggregate session stats (total sessions, avg health, total km)
 */
export const GetSessionsSummaryResponse = zod.object({
  totalSessions: zod.number(),
  avgHealthScore: zod.number(),
  totalDistanceKm: zod.number(),
  anomalyCount: zod.number(),
  recentSessions: zod.array(
    zod.object({
      id: zod.number(),
      vehicleId: zod.number(),
      startedAt: zod.coerce.date(),
      endedAt: zod.coerce.date().nullish(),
      maxRpm: zod.number().nullish(),
      maxSpeedKmh: zod.number().nullish(),
      avgHealthScore: zod.number().nullish(),
      distanceKm: zod.number().nullish(),
      anomalyDetected: zod.boolean(),
    }),
  ),
});

/**
 * @summary List diagnostic results
 */
export const ListDiagnosticsQueryParams = zod.object({
  vehicleId: zod.coerce.number().optional(),
});

export const ListDiagnosticsResponseItem = zod.object({
  id: zod.number(),
  vehicleId: zod.number().nullish(),
  dtcCode: zod.string(),
  severity: zod.enum(["low", "medium", "high", "critical"]),
  description: zod.string(),
  rootCauses: zod.array(zod.string()),
  repairSteps: zod.array(zod.string()),
  estimatedCostMin: zod.number().nullish(),
  estimatedCostMax: zod.number().nullish(),
  resolvedAt: zod.coerce.date().nullish(),
  createdAt: zod.coerce.date(),
});
export const ListDiagnosticsResponse = zod.array(ListDiagnosticsResponseItem);

/**
 * @summary Analyze a DTC fault code
 */
export const AnalyzeDtcBody = zod.object({
  dtcCode: zod.string(),
  vehicleId: zod.number().optional(),
});

export const AnalyzeDtcResponse = zod.object({
  id: zod.number(),
  vehicleId: zod.number().nullish(),
  dtcCode: zod.string(),
  severity: zod.enum(["low", "medium", "high", "critical"]),
  description: zod.string(),
  rootCauses: zod.array(zod.string()),
  repairSteps: zod.array(zod.string()),
  estimatedCostMin: zod.number().nullish(),
  estimatedCostMax: zod.number().nullish(),
  resolvedAt: zod.coerce.date().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get all component predictions for a vehicle
 */
export const ListPredictionsQueryParams = zod.object({
  vehicleId: zod.coerce.number(),
});

export const ListPredictionsResponseItem = zod.object({
  id: zod.number(),
  vehicleId: zod.number(),
  component: zod.string(),
  status: zod.enum(["good", "watch", "warning", "critical"]),
  kmUntilService: zod.number(),
  confidence: zod.number(),
  lastCheckedAt: zod.coerce.date(),
});
export const ListPredictionsResponse = zod.array(ListPredictionsResponseItem);

/**
 * @summary List all stock ECU maps in the vault
 */
export const ListMapsResponseItem = zod.object({
  id: zod.number(),
  vin: zod.string(),
  make: zod.string(),
  model: zod.string(),
  year: zod.number(),
  calId: zod.string(),
  description: zod.string().nullish(),
  checksumHex: zod.string().nullish(),
  sizeBytes: zod.number().nullish(),
  createdAt: zod.coerce.date(),
});
export const ListMapsResponse = zod.array(ListMapsResponseItem);

/**
 * @summary Register a new stock ECU map entry
 */
export const CreateMapBody = zod.object({
  vin: zod.string(),
  make: zod.string(),
  model: zod.string(),
  year: zod.number(),
  calId: zod.string(),
  description: zod.string().optional(),
  checksumHex: zod.string().optional(),
  sizeBytes: zod.number().optional(),
});

/**
 * @summary Get a specific ECU map
 */
export const GetMapParams = zod.object({
  id: zod.coerce.number(),
});

export const GetMapResponse = zod.object({
  id: zod.number(),
  vin: zod.string(),
  make: zod.string(),
  model: zod.string(),
  year: zod.number(),
  calId: zod.string(),
  description: zod.string().nullish(),
  checksumHex: zod.string().nullish(),
  sizeBytes: zod.number().nullish(),
  createdAt: zod.coerce.date(),
});