# lib/api-zod/src/generated/types/createSessionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeuronDrive API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSessionBody {
  vehicleId: number;
  maxRpm?: number;
  maxSpeedKmh?: number;
  avgHealthScore?: number;
  distanceKm?: number;
  anomalyDetected: boolean;
}