# lib/api-zod/src/generated/types/diagnostic.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeuronDrive API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DiagnosticSeverity } from "./diagnosticSeverity";

export interface Diagnostic {
  id: number;
  vehicleId?: number | null;
  dtcCode: string;
  severity: DiagnosticSeverity;
  description: string;
  rootCauses: string[];
  repairSteps: string[];
  estimatedCostMin?: number | null;
  estimatedCostMax?: number | null;
  resolvedAt?: Date | null;
  createdAt: Date;
}