# lib/api-zod/src/generated/types/ecuMap.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeuronDrive API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EcuMap {
  id: number;
  vin: string;
  make: string;
  model: string;
  year: number;
  calId: string;
  description?: string | null;
  checksumHex?: string | null;
  sizeBytes?: number | null;
  createdAt: Date;
}