# lib/api-zod/src/generated/types/prediction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeuronDrive API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PredictionStatus } from "./predictionStatus";

export interface Prediction {
  id: number;
  vehicleId: number;
  component: string;
  status: PredictionStatus;
  kmUntilService: number;
  confidence: number;
  lastCheckedAt: Date;
}