# lib/api-zod/src/generated/types/session.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeuronDrive API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Session {
  id: number;
  vehicleId: number;
  startedAt: Date;
  endedAt?: Date | null;
  maxRpm?: number | null;
  maxSpeedKmh?: number | null;
  avgHealthScore?: number | null;
  distanceKm?: number | null;
  anomalyDetected: boolean;
}