# lib/api-zod/src/generated/types/sessionSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeuronDrive API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Session } from "./session";

export interface SessionSummary {
  totalSessions: number;
  avgHealthScore: number;
  totalDistanceKm: number;
  anomalyCount: number;
  recentSessions: Session[];
}