# lib/api-zod/src/generated/types/vehicle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NeuronDrive API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Vehicle {
  id: number;
  vin: string;
  make: string;
  model: string;
  year: number;
  engineCode?: string;
  mileageKm: number;
  healthScore: number;
  createdAt: Date;
}